pyproject.toml
[project]
name = "elysiumdc"
version = "2.13"
description = "Elysium Discord Bot - A multi-purpose Discord bot"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.5.0,<3.0.0",
    "python-dotenv>=1.0.0",
    "yt-dlp>=2023.10.7",
    "requests>=2.31.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
